pyproject.toml
[tool.poetry]
name = "check-distributed-bot"
version = "0.1.0"
description = ""
authors = ["Andrew <veselov.andrey@rocknblock.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.2"
pydantic-settings = "^2.2.1"
web3 = "^6.16.0"
pyyaml = "^6.0.1"
uvloop = "^0.19.0"
redis = "^5.0.3"
aiogram = "^3.4.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
ptpython = "^3.0.26"
alembic = "^1.13.1"
celery = "^5.3.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
